Extract shared iOS 15 navbar styling
