+++ ZIGSIMPlus/Views/CommandSelectionViewController.swift
@@ -7,7 +7,6 @@
 //
 
 import MarkdownKit
-import SVProgressHUD
 import UIKit
 
 typealias CommandToSelect = (labelString: String, isAvailable: Bool)
@@ -96,18 +95,7 @@ final class CommandSelectionViewController: UIViewController {
     }
 
     private func adjustNavigationDesign() {
-        Utils.setTitleImage(navigationController!.navigationBar)
-        if #available(iOS 15.0, *) {
-            let appearance = UINavigationBarAppearance()
-            appearance.configureWithOpaqueBackground()
-            appearance.backgroundColor = Theme.dark
-            navigationController?.navigationBar.standardAppearance = appearance
-            navigationController?.navigationBar.scrollEdgeAppearance = appearance
-            navigationController?.navigationBar.tintColor = Theme.main
-        } else {
-            navigationController?.navigationBar.barTintColor = Theme.dark
-            navigationController?.navigationBar.tintColor = Theme.main
-        }
+        Utils.configureNavigationBar(navigationController!.navigationBar)
     }
 
     private func convertMessageFromStringToAttributedText(_ msg: String) -> NSMutableAttributedString {
